+++ components/table/filterDropdown.tsx
@@ -5,6 +5,10 @@ import Icon from '../icon';
 import Checkbox from '../checkbox';
 import Radio from '../radio';
 
+const FilterDropdownMenuWrapper = ({ onClick, children }) => (
+  <div className="ant-table-filter-dropdown" onClick={onClick}>{children}</div>
+);
+
 export default class FilterMenu extends React.Component {
   static defaultProps = {
     handleFilter() {},
@@ -113,7 +117,7 @@ export default class FilterMenu extends React.Component {
     const multiple = ('filterMultiple' in column) ? column.filterMultiple : true;
 
     const menus = column.filterDropdown ? column.filterDropdown : (
-      <div className="ant-table-filter-dropdown">
+      <FilterDropdownMenuWrapper>
         <Menu
           multiple={multiple}
           onClick={this.handleMenuItemClick}
@@ -138,7 +142,7 @@ export default class FilterMenu extends React.Component {
             {locale.filterReset}
           </a>
         </div>
-      </div>
+      </FilterDropdownMenuWrapper>
     );
 
     const dropdownSelectedClass = (this.props.selectedKeys.length > 0)
@@ -150,7 +154,6 @@ export default class FilterMenu extends React.Component {
         overlay={menus}
         visible={this.state.visible}
         onVisibleChange={this.onVisibleChange}
-        closeOnSelect={false}
       >
         <Icon title={locale.filterTitle} type="filter" className={dropdownSelectedClass} />
       </Dropdown>